extract email submission form for reuse
